fix: Fix dependencies with apps which use a more recent version of otel library
The package sharedotlptraces was providing a way to just pass additional opentelemetry attributes and let the library build the otel resource with it.
As convenient, this behavior force the usage of semconv library on a specific version, whereas more recent apps can use a more recent version of semconv.
Therefore, it constrains the app to use the same versions.
This refactoring remove this possibility of overriding otel resource by just providing additional attributes.
But an application can simply pass the resource to the fx container.